pyproject.toml
[project]
name = "sample-uv-app"
version = "0.1.0"
description = "LLM web search agent using Anthropic tool use with the ReAct pattern"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.84.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
ask = "main:app"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "strict"
